홍수명 커밋

+++ josso-final/src/main/webapp/WEB-INF/views/board/suggestionDetailPage.jsp
@@ -24,7 +24,6 @@
          <!-- Latest compiled JavaScript -->
          <script src="https://maxcdn.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js"></script>
          <!-- SmartEditor2 라이브러리(경로 맞춰줘야 함.) -->
-         <script type="text/javascript" src="../final_project/smartEditor/js/service/HuskyEZCreator.js" charset="utf-8"></script>
          <script>
             $(function(){
                 $('.side_title').click(function(){
@@ -77,17 +76,24 @@
 
             .left_td {
                 padding-bottom: 20px;
+                text-align:center;
                 width: 10%;
-                min-width: 100px;
-                background:lightgray;  
-                opacity:0.5;         
+                min-width: 200px;
+                background:#DCDCDC;
+                color:black; 
+                font-size: 1.2em;
             }
 
-
             .right_td {
+                padding-left: 30px;
                 padding-bottom: 20px;
-                width:50%;
-                padding-right: 30px;
+                font-weight:700;
+                font-size:1.2em;
+                min-width:300px;
+            }
+            
+            #right_td{
+                padding-bottom: 20px;
                 font-weight:700;
                 font-size:1.2em;
             }
@@ -128,32 +134,20 @@
            .row {
            		padding-top:30px;
            }
-
-
-     
+           
+           	#suggestion_title {
+				/* padding-left : 20px; */
+				/* background : #38A9BA; */
+				font-size : 2em;
+				color : #38A9BA;
+				width:280px;
+				font-weight : 700;
+			}
+			
+			td {
+				border : 1px solid lightgray;
+			}
         </style>
-
-<script type="text/javascript">
-    var oEditors = [];
-    $(function(){
-          nhn.husky.EZCreator.createInIFrame({
-              oAppRef: oEditors,
-              elPlaceHolder: "smartEditor", //textarea에서 지정한 id와 일치해야 함. 
-              //SmartEditor2Skin.html 파일이 존재하는 경로
-              sSkinURI: "smartEditor/SmartEditor2Skin.html",  
-              htParams : {
-                  // 툴바 사용 여부 (true:사용/ false:사용하지 않음)
-                  bUseToolbar : true,             
-                  // 입력창 크기 조절바 사용 여부 (true:사용/ false:사용하지 않음), 저는 크기 지정했기 땜에 사용 안 함.
-                  bUseVerticalResizer : false,     
-                  // 모드 탭(Editor | HTML | TEXT) 사용 여부 (true:사용/ false:사용하지 않음), 본에서도 쓰길래 true 해놓음,, 용도는 모르겠음
-                  bUseModeChanger : true,         
-              }, 
-              fCreator: "createSEditor2"
-          });
-        
-    });
-</script>
 
 <!--
 <script>
@@ -274,9 +268,9 @@
             </section>
         </div>
     </aside>
-    <nav >
+    <nav>
         <div class="nav_title">
-            <h2 id="suggestion_title" style="font-weight: 600;">건의사항 게시판</h2>
+            <h2 id="suggestion_title" style="font-weight: 600;">》 건의사항 게시판</h2>
         </div>
         <div class="nav_content">
             <p style="margin: 12px 17px; font-style: italic; ">임직원 여러분들의 의견을 무엇이든 건의할 수 있는 게시판입니다.<br>
@@ -284,50 +278,49 @@
             </p>
         </div>
     </nav>
-
-        
-        <main>
-            <form>
-            <div  style="padding-left:10px;">
-                <table style="width: 100%; text-align:center" id="board">
+        <main style="width:80%;">
+        <div style="width:100%">
+            <div  style="padding-left:20px; width:90%">
+                <table style="width: 100%; height:500px" id="board">
                     <thead>
                         <tr class="left_tr border-bottom"> 
-                            <th class="left_td" style="padding-top:20px">제목</th>
-                            <td class="right_td" style="padding-top:20px"><span>${suggestionBoard.boardTitle}</span></td><br>
+                            <th class="left_td" style="padding-top:20px;">제목</th>
+                            <td colspan="3" class="right_td" style="padding-top:20px"><span>${suggestionBoard.boardTitle}</span></td><br>
                         </tr>
                         <tr class="border-bottom" style="height: 50px;">
-                            <th class="left_td" style="padding-top: 20px;">파일첨부</th>
-                            <td>
+                            <th class="left_td" style="padding-top: 20px; width:5%">파일첨부</th>
+                            <td style="padding-left:30px; width:30%; color:gray">
                             <c:if test="${empty suggestionBoard.boardFile}">
 	                        	첨부파일 없음.
 	                        </c:if> <c:if test="${not empty suggestionBoard.boardFile}">
-									<a href="${pageContext.request.contextPath}/resources/multipartFile/${suggestionBoard.boardFile}"download> ${suggestionBoard.boardFile} </a>
+	                        	<a href="${pageContext.request.contextPath}/resources/multipartFile/${suggestionBoard.boardFile}"download>
+	                        		${suggestionBoard.boardFile}
+	                        	</a>
 							</c:if></td>
+                            <th class="left_td" style="padding-top:20px; width:5%">작성자</th>
+                            <td id="right_td" style="padding-top:20px; width:30%; text-align:center; font-size:1.2em"><span>${suggestionBoard.employeeName}</span></td><br>
                         </tr>
                     </thead>
                     <tbody>
                         <tr id="write_content" class="border-bottom">
                             <th class="left_td">내용</th>
-                            <td class="right_td"><div id="smartEditor" style="height: 400px; width: 100%;"><span>${suggestionBoard.boardContent}</span></div></td>
+                            <td colspan="3" class="right_td"><div id="smartEditor" style="height: 400px; width: 100%;"><span>${suggestionBoard.boardContent}</span></div></td>
                         </tr>
                     </tbody>    
                 </table>
             </div>
 
-           </form>
            <div class="row">        
-            	<div class="container col-sm-6" style="text-align:left; padding-left:30px">
+            	<div class="container col-sm-6" style="text-align:left; padding-left:40px">
                 	<button class="btn btn-info"><a href="list" style="color:white">목록</a></button>
             	</div>
-            	<div class="container col-sm-6" style="text-align:right">
+            	<div class="container col-sm-6" style="text-align:right; padding-right:11%">
 	                <button class="btn btn-info"><a href="updateBridge?num=${num}" style="color:white">수정하기</a></button>
-    	            <button class="btn btn-secondary"><a href="delete?num=${num}" style="color:white">삭제</a></button>
+    	            <button class="btn btn-secondary" onclick="confirm('정말 삭제하시겠습니까?')"><a href="delete?num=${num}" style="color:white">삭제</a></button>
             	</div>
             </div>
+           </div>
         </main>
-        <footer class="border-top">
-
-        </footer>
     </body>
 </html>
